requirements.txt
# Core ML stack
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Numerical / utilities
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0

# Vision–language + generative models
transformers>=4.40.0
datasets>=2.19.0
accelerate>=0.30.0
diffusers>=0.29.0
sentencepiece>=0.2.0
ftfy>=6.2.0
regex>=2023.12.25
open_clip_torch>=2.24.0

# Retrieval
faiss-cpu>=1.8.0

# Config / misc utils
PyYAML>=6.0.0
einops>=0.7.0
omegaconf>=2.3.0

# Logging / analysis / plots (optional but useful)
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
